BackEnd/requirements.txt
django==1.10.0
djangorestframework==3.7.7
django-rest-multiple-models==1.8.0
django-cors-headers==2.1.0